+++ frontend/components/commons/forms/questions-form/fields/text-area/TextArea.component.vue
@@ -18,7 +18,7 @@
     </div>
 
     <div class="container" :class="isFocused ? '--focused' : null">
-      <BaseRenderMarkdown
+      <BaseRenderMarkdownComponent
         v-if="useMarkdown && !isFocused"
         class="textarea--markdown"
         :markdown="value"
