+++ frontend/models/feedback-task-model/dataset-field/DatasetField.model.js
@@ -11,7 +11,7 @@ class DatasetField extends Model {
       order: this.number(0),
       title: this.string(""),
       is_required: this.boolean(false),
-      use_markdown: this.boolean(true), //TODO - remove true (only for testing)
+      use_markdown: this.boolean(true),
       component_type: this.string(null).nullable(),
     };
   }
